+++ src/components/Home/Footer/index.tsx
@@ -30,15 +30,6 @@ const Footer = () => {
             </Styled.Content>
           </Styled.AccountBox>
         </Styled.Info>
-        <Styled.Info>
-          <Styled.BoldContent>고객센터</Styled.BoldContent>
-          <Styled.Tel>1234-1234</Styled.Tel>
-          <Styled.Content>
-            평일 AM 09:00 ~ PM 05:30 / 주말, 공휴일 휴무
-          </Styled.Content>
-          <Styled.Content>점심 AM 12:00 ~ PM 01:00</Styled.Content>
-          <Button content="채팅상담" color="#285430" />
-        </Styled.Info>
       </Styled.InfoBox>
     </Styled.Wrapper>
   );
