+++ src/components/Order/InputField/index.tsx
@@ -1,6 +1,5 @@
 import {useState} from 'react'
-import theme from '@styles/theme';
-import styled from '@emotion/styled'
+import Styled from './style';
 import { useForm } from 'react-hook-form';
 import { emailOptions } from 'src/utils/register/emailListUtil';
 import {
@@ -11,7 +10,6 @@ import {
 import {
   requiredErrorMessage,
 } from 'src/utils/register/formUtil';
-import Option from '@assets/images/register/option.svg';
 import Button from './Button';
 
 interface InputFieldProps {
@@ -76,7 +74,7 @@ const InputField = ({label, required, field, width, placeholder, checkBoxLabel}:
               {...selectedEmail}
               placeholder="선택해주세요"
             />
-            <Styled.MoreBtn onClick={() => setIsDropdownOpen(prev => !prev)} />
+            {/* <div onClick={() => setIsDropdownOpen(prev => !prev)} />
             {isDropdownOpen && (
               <Styled.Dropdown>
                 {emailOptions.map((email, index) => (
@@ -88,7 +86,7 @@ const InputField = ({label, required, field, width, placeholder, checkBoxLabel}:
                   </Styled.Option>
                 ))}
               </Styled.Dropdown>
-            )}
+            )} */}
           </Styled.EmailOptionWrapper>
         </Styled.EmailWrapper>
       ) : field === 'phone' ? (
@@ -139,84 +137,4 @@ const InputField = ({label, required, field, width, placeholder, checkBoxLabel}:
 
 export default InputField
 
-const Styled = {
-  InputWrapper: styled.div<{field: string}>`
-    position: relative;
-    display: flex;
-    align-items: ${({field}) => field === 'coupon' || field === 'point' ? 'flex-start' : 'center'};
-  `,
-  Label: styled.label`
-    font-size: 16px;
-    width: 95px;
-  `,
-  AstBox: styled.span`
-    color: ${theme.colors.pointGreen};
-    margin: 2px;
-  `,
-  Input: styled.input<{width?:number}>`
-    outline: none;
-    width: ${({ width }) => width ? width: '300'}px;
-    height: 40px;
-    border-radius: 4px;
-    padding-left: 10px; 
-    border: 1px solid #ABABAB;
-  `,
-  EmailWrapper: styled.div`
-    display: flex;
-    align-items: center;
-    gap: 8px;
-  `,
-  EmailAt: styled.p`
-    font-size: 18px;
-  `,
-  EmailOptionWrapper: styled.div`
-    position: relative;
-  `,
-  MoreBtn: styled(Option)`
-    position: absolute;
-    top: 0;
-    right: 0;
-    cursor: pointer;
-    margin-right: 5px;
-  `,
-  Dropdown: styled.ul`
-    width: 300px;
-    position: absolute;
-    max-height: 130px;
-    overflow-y: auto;
-    border: 1px solid #000000;
-    background-color: #fff;
-    border-top: none;
-    padding: 0;
-    margin: 0;
-    list-style: none;
-    z-index: 1;
-  `,
-  Option: styled.li`
-      font-size: 14px;
-      padding: 10px;
-      cursor: pointer;
-      &:hover {
-        background-color: #f1f1f1;
-      }
-  `,
-  FlexWrapper: styled.div`
-    display: flex;
-  `,
-  FlexGapWrapper: styled.div`
-    display: flex;
-    align-items: center;
-    gap: 11px;
-  `,
-  FlexColumnWrapper: styled.div`
-    display: flex;
-    flex-direction: column;
-    gap: 17px;
-  `,
-  Explanation: styled.p`
-    padding: 6px 12px;
-    font-size: 16px;
-    border-radius: 5px;
-    background-color: #F0F0F0;
-  `,
-}
+
